requirements.txt
# Core
pillow~=11.2.1
bleak~=0.22.3
fastapi~=0.115.12
starlette~=0.46.2